requirements.txt
argparse=1.1
requests==2.22.0
pillow==7.0.0